errno moved to scene_parse, new header & other
